+++ src/presets/index.ts
@@ -7,7 +7,7 @@ export * from "./deno";
 export * from "./digital-ocean";
 export * from "./firebase";
 export * from "./heroku";
-export * from "./layer0";
+export * from "./edgio";
 export * from "./netlify";
 export * from "./nitro-dev";
 export * from "./nitro-prerender";
@@ -19,3 +19,4 @@ export * from "./service-worker";
 export * from "./stormkit";
 export * from "./vercel";
 export * from "./cleavr";
+export * from "./layer0";
